final/requirements.txt
numpy
Pillow
opencv-python
torch
chess
cairosvg